migrated everything implemented in users to profile_app
