+++ FitFriendsServer/settings.py
@@ -47,7 +47,7 @@ INSTALLED_APPS = [
     'allauth.socialaccount',
     'dj_rest_auth.registration',
     'allauth.socialaccount.providers.google',
-
+    'profile_app',
 ]
 
 MIDDLEWARE = [
