+++ profile_app/serializers.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from .models import Profile
+
+class ProfileSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Profile
+        fields = ('id', 'name', 'image')
