+++ users/admin.py
@@ -1,5 +1,4 @@
 from django.contrib import admin
-from .models import User, Profile
+from .models import User
 # Register your models here.
 admin.site.register(User)
-admin.site.register(Profile)
