+++ users/models.py
@@ -15,14 +15,3 @@ class User(AbstractUser):
 
     def __str__(self):
         return self.email
-#profile class
-class Profile(models.Model):
-     id = models.OneToOneField(User, on_delete=models.CASCADE)
-     username = models.CharField(max_length=100)
-     #friends list maybe?
-     #profile picture portion
-     profile_pic = models.ImageField(upload_to="images/")
-     #username is displayed 
-     def __str__(self):
-         return self.username
-#update profile function maybe
