+++ users/serializers.py
@@ -36,10 +36,4 @@
 #         Token = Token.objects.create(user=user)
 #         return user
 
-from rest_framework import serializers
-from .models import Profile
 
-class ProfileSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Profile
-        fields = ('id', 'name', 'image')
