+++ users/urls.py
@@ -1,19 +1,8 @@
 from django.contrib import admin
 from django.urls import include, path
 
-'''urlpatterns = [
-    path('register/', include('dj_rest_auth.registration.urls')),
-    path('',include('dj_rest_auth.urls'))
-]'''
-
-from .views import save_profile_data, change_image, change_name, return_profile
-
 urlpatterns = [
     path('register/', include('dj_rest_auth.registration.urls')),
-    path('',include('dj_rest_auth.urls')),
-    path('save-profile/', save_profile_data, name='save_profile'),
-    path('change-image/', change_image, name='change_image'),
-    path('change-name/', change_name, name='change_name'),
-    path('return-profile/', return_profile, name='return_profile')
+    path('',include('dj_rest_auth.urls'))
 ]
